+++ Source/projections/Builders/Copies/_exports.ts
@@ -1,0 +1,6 @@
+// Copyright (c) Dolittle. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+export { CopyToMongoDBBuilder } from './CopyToMongoDBBuilder';
+export { CopyToMongoDBCallback } from './CopyToMongoDBCallback';
+export { ICopyToMongoDBBuilder } from './ICopyToMongoDBBuilder';
